+++ src/pf_accum.h
@@ -28,7 +28,7 @@ struct pf_accum {
 
 struct accum_node {
     char *docno;
-    double val;
+    long double val;
     bool is_set;
     size_t count;
 };
@@ -40,7 +40,7 @@ void
 pf_accum_free(struct pf_accum *htable);
 
 unsigned long
-pf_accum_update(struct pf_accum **htable, const char *val, double score);
+pf_accum_update(struct pf_accum **htable, const char *val, long double score);
 
 struct pf_accum *
 pf_accum_rehash(struct pf_accum *htable);
